A start to some fixes
